feat: add ValidateButtonOpportunity component
