+++ packages/apps/crm/src/components/templates/index.js
@@ -24,3 +24,4 @@ export {default as LeadSearchBar} from './LeadSearchBar/LeadSearchBar';
 export {default as EventSearchBar} from './EventSearchBar/EventSearchBar';
 export {default as OpportunitySearchBar} from './OpportunitySearchBar/OpportunitySearchBar';
 export {default as ProspectSearchBar} from './ProspectSearchBar/ProspectSearchBar';
+export {default as ValidateButtonOpportunity} from './ValidateButtonOpportunity/ValidateButtonOpportunity';
